pyproject.toml
[project]
name = "rr-dtp"
version = "0.1.0"
description = ""
authors = [
    {name = "Chugunova Tatiana",email = "t.chugunova24@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "torch (>=2.7.0,<3.0.0)",
    "torchaudio (>=2.7.0,<3.0.0)",
    "torchvision (>=0.22.0,<0.23.0)",
    "numpy (>=2.2.6,<3.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "ultralytics (>=8.3.148,<9.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "opencv-python-headless (>=4.11.0.86,<5.0.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
